refactor: 주석 추가, goal, project, memoir dto에 있던 함수들 Assembler로 이동
