+++ src/main/java/com/letsRoll/letsRoll/Member/dto/res/MemberResDto.java
@@ -12,18 +12,4 @@ public class MemberResDto {
     private Long id;
     private String nickname;
     private String role;
-
-    public static List<MemberResDto> fromEntities(List<Member> members) {
-        return members.stream()
-                .map(MemberResDto::fromEntity)
-                .toList();
-    }
-
-    public static MemberResDto fromEntity(Member member) {
-        MemberResDto memberDto = new MemberResDto();
-        memberDto.setId(member.getId());
-        memberDto.setNickname(member.getNickname());
-        memberDto.setRole(member.getRole());
-        return memberDto;
-    }
 }
