+++ src/main/java/com/letsRoll/letsRoll/Memoir/service/MemoirService.java
@@ -3,6 +3,7 @@ package com.letsRoll.letsRoll.Memoir.service;
 import com.letsRoll.letsRoll.Goal.dto.res.GoalResDto;
 import com.letsRoll.letsRoll.Goal.entity.Goal;
 import com.letsRoll.letsRoll.Member.repository.MemberRepository;
+import com.letsRoll.letsRoll.Memoir.dto.MemoirAssembler;
 import com.letsRoll.letsRoll.Memoir.dto.res.MemoirResDto;
 import com.letsRoll.letsRoll.Memoir.entity.Memoir;
 import com.letsRoll.letsRoll.Memoir.repository.MemoirRepository;
@@ -23,13 +24,13 @@ public class MemoirService {
 
     public List<MemoirResDto> getMemoirs(Long projectId) {
         List<Memoir> memoirs = memoirRepository.findByProjectId(projectId);
-        return MemoirResDto.fromEntities(memoirs);
+        return MemoirAssembler.fromEntities(memoirs);
     }
 
     public MemoirResDto getMemoir(Long memoirId) {
         Memoir memoir = memoirRepository.findById(memoirId)
                 .orElseThrow(() -> new BaseException(BaseResponseCode.NOT_FOUND_MEMOIR));
-        return MemoirResDto.fromEntity(memoir);
+        return MemoirAssembler.fromEntity(memoir);
     }
 }
